后端功能性代码锁板

+++ src/main/java/com/hzx/service/FriendService.java
@@ -31,6 +31,7 @@ public class FriendService {
     }
 
     public boolean deleteFriend(int uid1, int uid2) {
+        if(uid2==1) return false;
         return friendMapper.deleteFriend(uid1, uid2);
     }
 }
